pyproject.toml
[project]
name = "does-confidence-calibration-improve-conformal-prediction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8.19"
dependencies = [
    "numpy==1.24.4",
    "python-dotenv>=1.0.1",
    "torch==2.4.1",
    "torchvision==0.19.1",
    "tqdm>=4.67.3",
]
